Some more algos.
